clang-tidy fixes

+++ hws/double_microscope/number.c
@@ -8,10 +8,10 @@ BinaryNumber decodeNumber(double number)
     bits = *(uint64_t*)(&number);
 
     result.sign = (bits >> 63) & 0x1;
-    result.exponent = ((bits >> 52) & 0x7FF) - 1023;
+    result.exponent = (int16_t) ((bits >> 52) & 0x7FF) - 1023;
     uint64_t mantissa = bits & 0xFFFFFFFFFFFFF;
     mantissa |= (1ULL << 52);
-    result.mantissa = mantissa / (double)(1ULL << 52);
+    result.mantissa = (double)mantissa / (double)(1ULL << 52);
 
     return result;
 }
